# packages/contracts/openapi/depsis.yaml
openapi: 3.1.0
info:
  title: DEPSIS API
  version: 0.1.0-phase1
  summary: Debian tabanlı çok kullanıcılı NAS platformu — Faz 1 yüzeyi
  description: |
    Bu belge tek doğruluk kaynağıdır. `apps/web` istemcisi buradan üretilir
    (`openapi-typescript`, ADR-0002); elle yazılmış istemci tipi bulunamaz.

    Kapsam: Faz 1 uçları. Konsol, uygulama kataloğu ve uzak erişim de burada — sırasıyla ADR-0018, ADR-0019 ve ADR-0020. Üçü de cihazın DIŞINDAKİ bir şeyi yönetiyor (bir kabuk, podman, zerotier-one); DEPSIS hiçbirini paketlemez, kuruluysa yönetir ve kurulu değilse 503 döner. Konsol ayrıcalıklı ajanda ÇALIŞMAZ: §2.2'nin serbest komut yasağı duruyor.
    Faz 2-3'e aittir ve kasıtlı olarak yoktur — tanımlanmamış bir uç, "yakında" diye
    duran boş bir uçtan iyidir.

    Sistem aracısı IPC şeması burada DEĞİLDİR. ADR-0006 gereği onu Rust sahiplenir ve
    schemars ile üretir: bir güven sınırının sözleşmesini, o sınırı uygulayan taraf
    tanımlamalıdır.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /me/mfa:
    delete:
      tags: [auth]
      summary: TOTP kaydını kaldır
      description: |
        Parola yeniden isteniyor. §0.5: güvenlik riski taşıyan bir işlem sessizce
        yapılamaz — ikinci faktörü kaldırmak, ele geçirilmiş bir oturumun yapmak isteyeceği ilk
        şeydir.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordConfirmation' }
      responses:
        '204':
          description: Kaldırıldı
        '401': { $ref: '#/components/responses/Problem' }
...
  /users:
    get:
      tags: [users]
      summary: Organizasyonun hesaplarını listele
      description: |
        Yalnız yöneticiler. 403 döner, 404 değil — bu uç noktanın varlığı sır değil, kimin
        çağırabileceği sır. Başka kiracının nesnesi olan uç noktalar 404 döner ve bu fark
        tutarsızlık değil, karar.
      responses:
        '200':
          description: Hesaplar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPage' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [users]
      summary: Hesap oluştur
      description: |
        Parola sunucuda Argon2id ile özetlenir; düz metin hiçbir yerde saklanmaz.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Oluşturuldu
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [users]
      summary: Görünen ad, rol veya devre dışı durumunu değiştir
      description: |
        Parola ve e-posta BURADA DEĞİL. Yöneticinin belirlediği bir parola, yöneticinin bildiği
        bir paroladır — her hesabı denetimde ayırt edilemeyecek şekilde taklit edilebilir kılar.
        E-posta değişikliği ise parola sıfırlamanın nereye gideceğini sessizce taşır.

        Bir organizasyon en az bir etkin yönetici tutmak zorunda; bunu veritabanı trigger'ı
        uygular, çünkü iki yöneticinin birbirini eşzamanlı düşürmesi uygulama seviyesinde
        sayılamaz. Son yöneticiyi düşürmek 409 döner.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: Güncellendi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /users/{id}/password-reset:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [users]
      summary: Bir hesap için tek kullanımlık parola sıfırlama bileti aç
      description: |
        Yalnız yöneticiler. Parolasını unutan bir kullanıcının üründen kurtarılmasının yolu.

        YÖNETİCİ PAROLAYI BELİRLEMİYOR ve `PATCH /users/{id}` bunun neden böyle olması
        gerektiğini yazıyor: "yöneticinin belirlediği bir parola, yöneticinin bildiği bir
        paroladır." Dönen şey elden verilecek tek kullanımlık bir jeton; parolayı kullanıcı
        kendisi seçiyor.

        Bu, taklidi imkânsız kılmıyor — bir NAS'ın yöneticisi zaten root eşdeğeri. Değiştirdiği
        şey, taklidin SESSİZ olmaktan çıkması: jeton tek kullanımlık, yani yönetici kullanırsa
        kullanıcının kendi denemesi başarısız olur ve kurban olayı öğrenir.

        Yöneticinin KENDİ parolası isteniyor. Bir oturum, birinin açık bırakılmış dizüstünü ödünç
        alan kişinin sahip olduğu şeydir ve bu uç, öyle bir dizüstünde bulunabilecek en kullanışlı
        şey: cihazdaki her hesaba açılan bir kapı.

        Jeton BİR KEZ gösteriliyor. Veritabanında yalnız SHA-256 özeti duruyor, yani bu yanıt
        değerin var olduğu tek yer; geri okuyabilecek ikinci bir uç yok.

        Kendine sıfırlama açmak 409: `POST /me/password` bunu zaten doğru yapıyor — mevcut parolayı
        istiyor ve mevcut oturumu ayakta bırakıyor.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  maxLength: 1024
                  description: Çağıran YÖNETİCİNİN kendi parolası. Hedef kullanıcının değil.
      responses:
        '201':
          description: Bilet açıldı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasswordResetTicket' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /me/password:
    post:
      tags: [me]
      summary: Kendi parolasını değiştir
      description: |
        Mevcut parola, oturum zaten varken bile isteniyor: bir oturum, birinin açık bırakılmış
        dizüstünü ödünç aldığında sahip olduğu şeydir, ve bu adım olmadan hesabın sahibini kalıcı
        olarak dışarıda bırakmak için gereken tek şeydir.

        Diğer BÜTÜN oturumlar iptal edilir. Parola değişikliği insanın kimlik bilgilerinin
        başkasında olduğunu düşündüğünde yaptığı şeydir; o birini oturumda bırakmak değişikliği
        anlamsız kılar. Mevcut oturum ayakta kalır.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '200':
          description: Değiştirildi
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PasswordChanged' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: depsis_session
      description: |
        Sunucu tarafı oturum. JWT kullanılmaz — güvenlik olayında tüm oturumların iptal
        edilebilmesi gerekiyor ve stateless token bunu ucuz yapmıyor (ADR-0009).
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail:
          type: string
          description: Secret, dosya içeriği veya çağıranın görmeye yetkili olmadığı yol içermez.
        instance: { type: string }
        code:
          type: string
          description: Kapalı küme. Kaynak tanım packages/contracts/src/errors.ts.
        correlationId:
          type: string
          description: Her yanıtta ve aynı isteğin her log satırında bulunur (§14).
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemFieldError' }
        retryAfter: { type: integer, minimum: 0 }
...
    PasswordResetTicket:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
          description: |
            Tek kullanımlık. Bir daha gösterilmiyor — sunucuda yalnız SHA-256 özeti duruyor.
        expiresAt: { type: string, format: date-time }
    ProblemFieldError:
      type: object
      required: [pointer, code, message]
      properties:
        pointer: { type: string }
        code: { type: string }
        message: { type: string }
...
    PasswordConfirmation:
      type: object
      required: [password]
      description: |
        §0.5 gereği yeniden kimlik doğrulama. Oturum çerezi kimin olduğunu söyler; bu, o kişinin
        hâlâ klavyenin başında olduğunu söyler.
      properties:
        password: { type: string, maxLength: 1024 }
...
    User:
      type: object
      required: [id, username, role, disabled, createdAt]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: [string, 'null'] }
        role:
          type: string
          enum: [admin, member]
        disabled: { type: boolean }
        createdAt: { type: string, format: date-time }
    UserPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/User' }
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$'
          description: |
            Giriş kimliği. '@' içeremez — bir kullanıcı adının bir adresle karıştırılabilmesi,
            ne insan ne de gelecekteki bir ayrıştırıcı için iyi bir fikir.
        role:
          type: string
          enum: [admin, member]
          default: member
        password:
          type: string
          minLength: 12
          description: |
            Uzunluk tabanı, kompozisyon kuralı değil. "Bir rakam, bir sembol" kuralı insanları
            ölçülebilir biçimde `Passw0rd!`e itiyor ve dört karakter daha uzunluktan az kazandırıyor.
    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        role:
          type: string
          enum: [admin, member]
        disabled: { type: boolean }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword: { type: string, minLength: 12 }
    PasswordChanged:
      type: object
      required: [status, otherSessionsRevoked]
      properties:
        status: { type: string, enum: [ok] }
        otherSessionsRevoked: { type: integer }
...